Backend/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.32.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Flask-CORS==6.0.1
